edited max speed in location history to default to float 3.0

+++ app/Http/Controllers/CarController.php
@@ -121,7 +121,7 @@ class CarController extends Controller
             'car_name' => $car->car_name,
             'date' => $date,
             'total_distance' => round($totalDistance, 2),
-            'max_speed' => (double)($locations->max('speed') ?? 0), // Use (double) cast            'locations_count' => $locations->count(),
+            'max_speed' => (float)($locations->max('speed') ?? 0.0), // Use (double) cast            'locations_count' => $locations->count(),
             'locations' => GpsDataResource::collection($locations)
         ]);
     }
